Update UpdateUserTeamPlayerIsPlayForGameweek to take player names as well as keys and blank previous week's player's IsPlay column
Add sp to Update the IsCaptain by name or key

Correction to calculated column SeasonPart on DimGameweek

Add LastUpdated column to PlayerPointsPerGame tables, and add foreign keys to these tables

Added new sp:

GetUserTeamPlayerPointsForGameweek so cn get a report on player points for a single gameweek.

GetPlayerHistoryAgainstOpponentTeam to be able to get a players record against a certain team

GetAllPlayerPPGForSeasonHalfAndSeasonPart to get player points for SeasonHalf and SeasonPart to be able to assess future players.

GetAllPlayerPPGForHomeAgainstAwayMatches to be able to look at a players record in home games against away games.

Correction to unit test:
GetAllPlayerPPGForSeasonPrevious5Previous10AndAllGameweeks - @MaxCost added